@@ -1,9 +1,9 @@
 package service;
 
-import java.util.Scanner;
-
 import dto.User;
 import repository.UserRepository;
+
+import java.util.Scanner;
 
 
 /**
@@ -16,93 +16,95 @@ import repository.UserRepository;
  */
 public class UserOperations {
 
-	private final UserRepository repository = new UserRepository();
+    private final UserRepository repository = new UserRepository();
 
-	public void showUsersData() {
-		repository.getAll().forEach(System.out::println);
-	}
+    public void showUsersData() {
+        repository.getAll().forEach(System.out::println);
+    }
 
-	public User getUserByTabnum(int tabnum) {
-		return repository.getUserByTabnum(tabnum);
-	}
-	
-	public User getUserById(int id) {
-		return repository.getUserById(id);
-	}
+    public User getUserByTabnum(int tabnum) {
+        return repository.getUserByTabnum(tabnum);
+    }
 
-	public void createUser(User user) {
-		repository.create(user);
-		User createdUser = repository.getUserByTabnum(user.getTabnum());
-		if (createdUser == null) {
-			System.out.println("\nЧто-то пошло не так и пользователя не удалось создать: " + user);
-		} else {
-			System.out.println("\nДобавлен новый пользователь: " + createdUser);
-		}
-	}
-	public void updateUser(User user, User userFromDb) {
-		userFromDb = repository.getUserById(user.getId());
-		prepareUserForUpdate(user, userFromDb);
-		repository.update(userFromDb);
-		User updateUser = repository.getUserById(user.getId());
-		if (updateUser == null) {
-			System.out.println("\nЧто-то пошло не так, данные пользователя не удалось обновить: " + user);
-		} else {
-			System.out.println("\nОбновлены данные пользователя: " + updateUser);
-		}
-	}
-	
-	 public User prepareUserForUpdate(User user, User userFromDb) {
-		 if(userFromDb == null) {
-				throw new IllegalArgumentException("Ошибка обновления данных. Не удалось получить данные о пользователе " + userFromDb);
-			}
-			if(user.getTabnum() != 0) {
-				userFromDb.setTabnum(user.getTabnum());
-			}
-			if(user.getName() != null) {
-				userFromDb.setName(user.getName());
-			}
-			if(user.getSurname() != null) {
-				userFromDb.setSurname(user.getSurname());
-			}
-			if(user.getBirth() != null) {
-				userFromDb.setBirth(user.getBirth());
-			}
-			return userFromDb;
-	 }
-	
-	public void showUser(User user) {
-		repository.show(user);
-		User showUser = repository.getUserById(user.getId());
-		if (showUser == null) {
-			System.out.println("\nЧто-то пошло не так и пользователя не удалось показать: " + user);
-		} else {
-			System.out.println("\nПользователь: " + showUser);
-		}
-	}
-	public void deleteUser(User user) {
-		User showUser = repository.getUserById(user.getId());
-		System.out.println("\n" + showUser);
-		while(true) {
- 		System.out.println("\nПодтвердите удаление пользователя (yes/no): ");
-		Scanner scanner = new Scanner(System.in);
-			String response = scanner.nextLine().trim();
-			if("yes".equals(response)) {
-				repository.delete(user);
-				User deleteUser = repository.getUserById(user.getId());
-				if (deleteUser == null) {
-					System.out.println("\nПользователь удален." );
-					break;
-				} else {
-					System.out.println("\nЧто-то пошло не так и пользователя не удалось удалить.");
-				}	
-			} else if("no".equals(response)) {
-				System.out.println("Отмена удаления пользователя.");
-				break;
-			} else {
-				System.out.println("Ошибка. Введите yes или no!");
-				continue;
-			}
-			scanner.close();
-		}
-	}
+    public User getUserById(int id) {
+        return repository.getUserById(id);
+    }
+
+    public void createUser(User user) {
+        repository.create(user);
+        User createdUser = repository.getUserByTabnum(user.getTabnum());
+        if (createdUser == null) {
+            System.out.println("\nЧто-то пошло не так и пользователя не удалось создать: " + user);
+        } else {
+            System.out.println("\nДобавлен новый пользователь: " + createdUser);
+        }
+    }
+
+    public void updateUser(User user, User userFromDb) {
+        userFromDb = repository.getUserById(user.getId());
+        prepareUserForUpdate(user, userFromDb);
+        repository.update(userFromDb);
+        User updateUser = repository.getUserById(user.getId());
+        if (updateUser == null) {
+            System.out.println("\nЧто-то пошло не так, данные пользователя не удалось обновить: " + user);
+        } else {
+            System.out.println("\nОбновлены данные пользователя: " + updateUser);
+        }
+    }
+
+    public User prepareUserForUpdate(User user, User userFromDb) {
+        if (userFromDb == null) {
+            throw new IllegalArgumentException("Ошибка обновления данных. Не удалось получить данные о пользователе " + userFromDb);
+        }
+        if (user.getTabnum() != 0) {
+            userFromDb.setTabnum(user.getTabnum());
+        }
+        if (user.getName() != null) {
+            userFromDb.setName(user.getName());
+        }
+        if (user.getSurname() != null) {
+            userFromDb.setSurname(user.getSurname());
+        }
+        if (user.getBirth() != null) {
+            userFromDb.setBirth(user.getBirth());
+        }
+        return userFromDb;
+    }
+
+    public void showUser(User user) {
+        repository.show(user);
+        User showUser = repository.getUserById(user.getId());
+        if (showUser == null) {
+            System.out.println("\nЧто-то пошло не так и пользователя не удалось показать: " + user);
+        } else {
+            System.out.println("\nПользователь: " + showUser);
+        }
+    }
+
+    public void deleteUser(User user) {
+        User showUser = repository.getUserById(user.getId());
+        System.out.println("\n" + showUser);
+        while (true) {
+            System.out.println("\nПодтвердите удаление пользователя (yes/no): ");
+            Scanner scanner = new Scanner(System.in);
+            String response = scanner.nextLine().trim();
+            if ("yes".equals(response)) {
+                repository.delete(user);
+                User deleteUser = repository.getUserById(user.getId());
+                if (deleteUser == null) {
+                    System.out.println("\nПользователь удален.");
+                    break;
+                } else {
+                    System.out.println("\nЧто-то пошло не так и пользователя не удалось удалить.");
+                }
+            } else if ("no".equals(response)) {
+                System.out.println("Отмена удаления пользователя.");
+                break;
+            } else {
+                System.out.println("Ошибка. Введите yes или no!");
+                continue;
+            }
+            scanner.close();
+        }
+    }
 }
